Refacto load/export pipeline
